+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - **Internals** (**Note:** Skip this if you haven't created custon Elements inherrited from WUP...)
   - Added auto-mapping between attributes <=> options based on key-values in `$defaults`
   - Removed interface `Defaults`. Merged with interface `Options` and now contains all fields as required
-  - Now removing attributes always rollbacks to value defined in `$defaults`
+  - Now removing attributes/options always rollbacks to value defined in `$defaults`
 - [SpinElement](src/spinElement.ts). Defaults `fit` & `overflowTarget` = `auto` instead of `null`
 
 ## 0.8.1 (Aug 29, 2023)